pyproject.toml
[tool.poetry]
name = "stock-notifier"
version = "0.1.0"
description = ""
authors = ["Julius Frost <33183774+juliusfrost@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "stock_notifier"}]

[tool.poetry.dependencies]
python = "^3.10"
py-cord = "^2.4.1"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.5"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[tool.poetry.scripts]
run_bot = "stock_notifier.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
